LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/unregister-then-register-new-script.https-expected.txt



PASS Registering a new script URL while an unregistered registration is in use 
FAIL Registering a new script URL that 404s does not resurrect an unregistered registration assert_unreached: unexpected rejection: assert_unreached: register should reject the promise Reached unreachable code Reached unreachable code
FAIL Registering a new script URL that fails to install does not resurrect an unregistered registration assert_unreached: unexpected rejection: assert_equals: document should not load with a controller expected null but got object "[object ServiceWorker]" Reached unreachable code
